Refactor header to use server-side authentication and improve separation of concerns
